Fixed handling of PLAYER_SPECIALIZATION_CHANGED event
Adjusted the handling of the PLAYER_SPECIALIZATION_CHANGED event to ensure that the specialization information is retrieved accurately for the player. Added a short delay before fetching the specialization to allow the event to settle, ensuring stability in the retrieved data. This fix addresses issues where the event was triggered prematurely or with incomplete data, resulting in incorrect specialization information being used for updating the guild note.